Fix build for clang
Remove declaring _POSIX_SOURCE manually. Let the compiler take care of
this on its own, except on Windows where the compiler doesn't take care
of this. Additionally, include a missing header file for the definition
of struct sockaddr_in, which we need before we can allocate the struct.
With these fixes, we can build with Clang 18.1.5, although there are
many warnings.